Integration: pin a v5 version of the v5 module, and a size is int64
`require .../sdk-go/v5 v4.0.0` is not a go.mod Go will parse, and the
runner only moves the pin to the newest tag AFTER parsing, so the suite
failed before testing anything. v5.0.0 is the floor the range check reads.
Against 5.0.0 a published file's size is int64, which the suite still
returned as int.

+++ integration/database_test.go
@@ -273,7 +273,7 @@ func transferred(t *testing.T) *transfer {
 	return shared
 }
 
-func publishedSize(t *testing.T, meta *vpndetection.DatabaseMetadata) int {
+func publishedSize(t *testing.T, meta *vpndetection.DatabaseMetadata) int64 {
 	t.Helper()
 	if meta.Size == nil {
 		t.Fatalf("%s publishes no size to check a transfer against", datasetID)
